refactor(update): preserve run facts and separate runtime validation
Keep phase history and notice custody through bounded diagnostic eviction, close interrupted steps at terminalization, and move unchanged validation out of browser imports. No schema or stored-shape change.

(cherry picked from commit 03b41549c8db17757ac0a0ef4e9c57f4a7c9d1f1)
